+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/MainActivity.kt
@@ -1189,7 +1189,7 @@ class MainActivity : SimpleActivity(), DirectoryOperationsListener {
         if (currAdapter == null || forceRecreate) {
             initZoomListener()
             val fastscroller = if (config.scrollHorizontally) directories_horizontal_fastscroller else directories_vertical_fastscroller
-            DirectoryAdapter(this, dirsToShow, this, directories_grid, isPickIntent(intent) || isGetAnyContentIntent(intent), fastscroller) {
+            DirectoryAdapter(this, dirsToShow, this, directories_grid, isPickIntent(intent) || isGetAnyContentIntent(intent), directories_refresh_layout, fastscroller) {
                 val clickedDir = it as Directory
                 val path = clickedDir.path
                 if (clickedDir.subfoldersCount == 1 || !config.groupDirectSubfolders) {
@@ -1206,6 +1206,10 @@ class MainActivity : SimpleActivity(), DirectoryOperationsListener {
                 runOnUiThread {
                     directories_grid.adapter = this
                     setupScrollDirection()
+
+                    if (config.viewTypeFolders == VIEW_TYPE_LIST) {
+                        directories_grid.scheduleLayoutAnimation()
+                    }
                 }
             }
             measureRecyclerViewContent(dirsToShow)