+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/PanoramaPhotoActivity.kt
@@ -57,6 +57,12 @@ open class PanoramaPhotoActivity : SimpleActivity() {
         }
 
         window.statusBarColor = resources.getColor(R.color.circle_black_background)
+
+        if (config.maxBrightness) {
+            val attributes = window.attributes
+            attributes.screenBrightness = 1f
+            window.attributes = attributes
+        }
     }
 
     override fun onPause() {