+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/extensions/Context.kt
@@ -8,6 +8,7 @@ import android.database.Cursor
 import android.graphics.Bitmap
 import android.graphics.drawable.PictureDrawable
 import android.media.AudioManager
+import android.os.Process
 import android.provider.MediaStore.Files
 import android.provider.MediaStore.Images
 import android.widget.ImageView
@@ -113,6 +114,18 @@ fun Context.getSortedDirectories(source: ArrayList<Directory>): ArrayList<Direct
     if (sorting and SORT_BY_RANDOM != 0) {
         dirs.shuffle()
         return movePinnedDirectoriesToFront(dirs)
+    } else if (sorting and SORT_BY_CUSTOM != 0) {
+        val newDirsOrdered = ArrayList<Directory>()
+        config.customFoldersOrder.split("|||").forEach { path ->
+            val index = dirs.indexOfFirst { it.path == path }
+            if (index != -1) {
+                val dir = dirs.removeAt(index)
+                newDirsOrdered.add(dir)
+            }
+        }
+
+        dirs.mapTo(newDirsOrdered, { it })
+        return newDirsOrdered
     }
 
     dirs.sortWith(Comparator { o1, o2 ->
@@ -587,6 +600,11 @@ fun Context.tryLoadingWithPicasso(path: String, view: MySquareImageView, cropThu
 
 fun Context.getCachedDirectories(getVideosOnly: Boolean = false, getImagesOnly: Boolean = false, forceShowHidden: Boolean = false, callback: (ArrayList<Directory>) -> Unit) {
     ensureBackgroundThread {
+        try {
+            Process.setThreadPriority(Process.THREAD_PRIORITY_MORE_FAVORABLE)
+        } catch (ignored: Exception) {
+        }
+
         val directories = try {
             directoryDao.getAll() as ArrayList<Directory>
         } catch (e: Exception) {
@@ -812,7 +830,7 @@ fun Context.deleteMediumWithPath(path: String) {
 }
 
 fun Context.updateWidgets() {
-    val widgetIDs = AppWidgetManager.getInstance(applicationContext).getAppWidgetIds(ComponentName(applicationContext, MyWidgetProvider::class.java))
+    val widgetIDs = AppWidgetManager.getInstance(applicationContext)?.getAppWidgetIds(ComponentName(applicationContext, MyWidgetProvider::class.java)) ?: return
     if (widgetIDs.isNotEmpty()) {
         Intent(applicationContext, MyWidgetProvider::class.java).apply {
             action = AppWidgetManager.ACTION_APPWIDGET_UPDATE