+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/helpers/Constants.kt
@@ -88,6 +88,7 @@ const val FOLDER_MEDIA_COUNT = "folder_media_count"
 const val LIMIT_FOLDER_TITLE = "folder_limit_title"
 const val THUMBNAIL_SPACING = "thumbnail_spacing"
 const val FILE_ROUNDED_CORNERS = "file_rounded_corners"
+const val CUSTOM_FOLDERS_ORDER = "custom_folders_order"
 
 // slideshow
 const val SLIDESHOW_INTERVAL = "slideshow_interval"